tools/requirements.txt
Pillow==11.1.0
numpy
